This is a Testing platform software similar to software used in most of Competative Exams with time constaint. Java swing is used.
THis project is based on this concept and this is  simple Online
platform in which the candidate is provided with the instructions to
attempt the 20 multiple choice questions. Candidates can navigate to any
other question at any time. Also the candidate can submit the paper at
any time and get his results. We also enforced strict time constraints
and automatic submission at the end of given time.
In this project Java swing is used through which  many classes like
Action Listener and Event Handler are applied.